pyproject.toml
[project]
name = "unk029-bank-app"
version = "1.0.0"
description = "UNK029 Banking Application - FastAPI backend with MCP server and ADK agent"
authors = [
    { name = "UNK029 Team", email = "unk029@openconsultinguk.com" }]
license = "MIT"
license-files = ["LICENSE*"]
keywords = ["banking", "fastapi", "mcp", "agent", "adk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
    "Topic :: Office/Business :: Financial",
]
requires-python = ">=3.10,<3.13"
dependencies = [
    "unk029>=0.1.3",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "fastmcp>=0.1.0",
    "google-adk[web]>=0.5.0",
    "httpx>=0.25.0",
    "rich",
]

[project.optional-dependencies]
# Development/testing extras
dev = [
    "pytest",
    "pytest-cov",
]

[project.urls]
Homepage = "https://github.com/unk029-openconsultinguk/unk029_bank_app"
Documentation = "https://github.com/unk029-openconsultinguk/unk029_bank_app"
Repository = "https://github.com/unk029-openconsultinguk/unk029_bank_app"

#[project.scripts]
#unk029 = "unk029.cli:app"

[dependency-groups]
nox = [
    "nox-uv",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-randomly",
]
type_check = [
    "mypy",
    # Add "types-" stub packages as needed: https://github.com/python/typeshed/tree/main/stubs
]
lint = [
    "ruff",
]
docs = [
    "mkdocs-material",
    "mkdocs-htmlproofer-plugin",

    # Python API documentation (not neccessary for applications).
    "mkdocstrings[python]",
    # Autodoc.
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
]
licenses = [
    "pip-licenses-cli",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bank_app"]


# For non-src directory projects, explicitly set top level model name:
#[tool.uv.build-backend]
#module-name = "my-app"
#module-root = ""

[tool.uv]
default-groups = "all"
# Non-package applications can uncomment the following line and delete [build-system] section to
# avoid being built or installed: https://docs.astral.sh/uv/reference/settings/#package
package = false

[tool.mypy]
ignore_missing_imports = true
strict = true
enable_error_code = ["deprecated", "exhaustive-match", "explicit-override"]
# If certain strict config options are too pedantic for a project,
# disable them selectively here by setting to false.

[tool.ruff]
line-length = 99
src = ["src"]
# Ruff will automatically exclude all files listed in .gitignore as well as common temporary Python
# tool directories.
# To exclude additional folders, use extend-exclude.

[tool.ruff.lint]
select = [
    "F", # pyflakes
    "E", # pycodestyle
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "RUF", # ruff
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "PIE", # flake8-pie
    "PT", # flake-pytest-style
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
]
extend-ignore = [
    "RUF005",
    "RUF012",
]
unfixable = [
    # Disable removing unused imports by default and only enable within nox so editors don't delete
    # unused imports while the user is in the middle of editing a file on save.
    "F401",
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = false
# For non-src directory projects, explicitly set top level package names:
# known-first-party = ["my-app"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.pytest.ini_options]
addopts = [
    "--strict-config",
    "--strict-markers",
]
xfail_strict = true
filterwarnings = [
    # When running tests, treat warnings as errors (e.g. -Werror).
    # See: https://docs.pytest.org/en/latest/reference/reference.html#confval-filterwarnings
    "error",
    # Add additional warning suppressions as needed here. For example, if a third-party library
    # is throwing a deprecation warning that needs to be fixed upstream:
    # "ignore::DeprecationWarning:typer",
]

[tool.coverage.run]
branch = true

# To globally exclude additional code blocks from code coverage reporting, see:
# https://coverage.readthedocs.io/en/latest/excluding.html#advanced-exclusion
[tool.coverage.report]
exclude_also = [
    "case _ as unreachable",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]

[[tool.uv.index]]
name = "nexus"
url = "https://nexus.openconsultinguk.com/repository/sandbox/simple/"

[tool.uv.sources]
unk029 = { index = "nexus" }

